requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
gunicorn==21.2.0
numpy==1.23.5
tensorflow==2.11.0
pillow==10.1.0
python-multipart==0.0.6
